Update web3 version
Update to resolve web3-core-requestmanager issue reported in https://github.com/ChainSafe/web3.js/pull/3721

Assume a provider is passed to DydxClient, where the request function uses the 'this' keyword (for example to ensure initialization of the provider, perhaps fetching some data).
When calling functions, such as:
`client.eth.exchange.deposit(....)`

The flow will eventually reach:
'RequestManager.send'
In this, when calling the provider.request, we do not bind the provider, thus the value of 'this' is undefined, and any call to 'this' will result in an error due to undefined value.

The fix will add the '.bind(this.provider)' which is currently missing, and make the 'this' value to be defined to the provider initially provided to the client.